requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
xarray>=2024.1.0
netCDF4>=1.6.5
scipy>=1.12.0
scikit-learn>=1.4.0
cartopy>=0.22.0
cfgrib>=0.9.11.0
eccodes>=1.6.1
pyproj>=3.6.1
requests>=2.31.0

# Google ADK dependencies
google-adk>=0.1.0  # Required by our package
google-cloud-aiplatform[agent-engines]>=1.95.1
google-cloud-core>=2.4.0
google-api-core>=2.17.0
google-auth>=2.27.0
google-cloud-storage>=2.14.0
google-generativeai>=0.3.0  # For Gemini models
vertexai>=0.0.1  # For Vertex AI integration

# Development dependencies
pylint>=3.1.0
black>=23.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
mypy>=1.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
aiohttp>=3.9.0

# New dependencies
fastapi>=0.68.0
jinja2
uvicorn>=0.15.0
pydantic>=2.0.0
python-dotenv>=0.19.0
requests>=2.26.0
aiohttp>=3.8.0
asyncio>=3.4.3
typing-extensions>=4.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0

# Testing
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-xdist>=2.5.0
coverage>=6.3.0
graphviz>=0.19.0

# Visualization and Reporting
networkx>=3.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0

# Development
black>=21.12b0
isort>=5.12.0
flake8>=4.0.0
mypy>=0.910
pre-commit>=2.17.0

# Documentation
mkdocs>=1.2.0
mkdocs-material>=8.0.0
mkdocstrings>=0.18.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5

# Monitoring and Logging
prometheus-client>=0.12.0
python-json-logger>=2.0.0
structlog>=21.5.0

# Utilities
python-dateutil>=2.8.2
pytz>=2021.3
tenacity>=8.0.0
cachetools>=5.0.0

# Additional dependencies
jsonschema>=4.0.0
great-expectations>=0.18.0
polars>=0.20.0
pyarrow>=14.0.0
fastparquet>=2023.0.0
ruff>=0.1.0

# Google Cloud dependencies
google-cloud-bigquery>=3.0.0
google-cloud-storage>=2.0.0
google-cloud-firestore>=2.0.0
google-cloud-pubsub>=2.0.0
google-api-core>=2.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.0.0

# Development
black>=23.0.0
isort>=5.0.0
flake8>=6.0.0
mypy>=1.0.0

# Testing and development
pytest==8.4.0
pytest-cov==6.2.1
pytest-asyncio==1.0.0
pytest-mock==3.14.1
pytest-benchmark==5.1.0

# Performance testing and optimization (Phase 5)
locust==2.17.0
psutil==5.9.6
memory-profiler==0.61.0
py-spy==0.3.14
redis==5.0.1
cachetools==5.3.2

# Security and monitoring (Phase 5)
cryptography==41.0.8
pyjwt==2.8.0
bcrypt==4.1.2
prometheus-client==0.19.0
structlog==23.2.0

# Data processing
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3 